move LogMessageAdapter to home directory
