+++ sampleapp/src/main/java/io/matthewnelson/sampleapp/topl_android/MyEventBroadcaster.kt
@@ -68,7 +68,7 @@ package io.matthewnelson.sampleapp.topl_android
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import io.matthewnelson.sampleapp.ui.LogMessageAdapter
+import io.matthewnelson.sampleapp.ui.fragments.home.LogMessageAdapter
 import io.matthewnelson.topl_core_base.BaseConsts
 import io.matthewnelson.topl_service.service.components.onionproxy.TorServiceEventBroadcaster
 import io.matthewnelson.topl_service.util.ServiceUtilities
